#651: Moved the test cases related to RqMultipartSmart into a dedicated TestCase
