+++ src/main/java/org/takes/rq/multipart/RqMultipartBase.java
@@ -72,7 +72,7 @@ import org.takes.rq.TempInputStream;
  *
  * @author Yegor Bugayenko (yegor@teamed.io)
  * @version $Id$
- * @since 0.32.8
+ * @since 0.33
  * @see <a href="http://www.w3.org/TR/html401/interact/forms.html">
  *  Forms in HTML</a>
  * @checkstyle ClassDataAbstractionCouplingCheck (500 lines)
