+++ src/main/java/org/takes/rq/multipart/RqMultipartFake.java
@@ -37,7 +37,7 @@ import org.takes.rq.RqWithHeaders;
  * Fake decorator.
  * @author Nicolas Filotto (nicolas.filotto@gmail.com)
  * @version $Id$
- * @since 0.32.8
+ * @since 0.33
  */
 public final class RqMultipartFake implements RqMultipart {
     /**
