+++ src/main/java/org/takes/rq/multipart/RqMultipartSmart.java
@@ -35,7 +35,7 @@ import org.takes.rq.RqMultipart;
  * Smart decorator.
  * @author Nicolas Filotto (nicolas.filotto@gmail.com)
  * @version $Id$
- * @since 0.32.8
+ * @since 0.33
  */
 public final class RqMultipartSmart implements RqMultipart {
     /**
