+++ src/test/java/org/takes/rq/multipart/package-info.java
@@ -27,6 +27,6 @@
  *
  * @author Nicolas Filotto (nicolas.filotto@gmail.com)
  * @version $Id$
- * @since 0.32.8
+ * @since 0.33
  */
 package org.takes.rq.multipart;
